pyproject.toml
[project]
name = "agents-maintenance-dashboard"
version = "1.0.3"
description = "Agent maintenance dashboard for Civic Interconnect"
authors = [
  { name = "Civic Interconnect", email = "info@civic-interconnect.org" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "requests",
  "PyYAML",
  "typer[all]",
  "rich",
  "civic-lib-core @ git+https://github.com/civic-interconnect/civic-lib-core.git@main",
]

[project.optional-dependencies]
dev = [
  "pre-commit",
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "ruff"
]

[project.urls]
"Homepage" = "https://civic-interconnect.github.io/agents-maintenance-dashboard/"
"Source" = "https://github.com/civic-interconnect/agents-maintenance-dashboard"

[project.scripts]
app-agents = "app_agents.cli.cli:app"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.ruff]
line-length = 100
target-version = "py312"
fix = true
unsafe-fixes = true
preview = true

[tool.ruff.lint]
extend-select = ["E", "F", "W", "I", "UP", "C90", "N", "B", "A", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
docstring-code-format = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra --strict-markers"
testpaths = [
  "tests"
]
python_files = "test_*.py"
